Add CompositeCollector for non-concurrent shard-collect
This adds a CompositeCollector which controls 1+ sub-collectors and
launches them sequentially.

There are two reasons for this:

1. Performance:

Our disk access is mostly done lazy.
This means execution looked mostly like this:

    S1      S2
     \     /
      \   /
    synchronized
     RowMerger
       |
      RR (disk IO)

The RowMerger effectively eliminates any multi-threading benefits here,
so using threaded execution is just overhead.

The exceptions are projections with shard level granularity. These will
still be run as before because there computations and disk IO happen
before the merge:

    S1    S2
     |    |
     P    P
     \   /
      \ /
      Merger
        |
       RR

    Simple benchmark result:

        Statement: select * from uservisits limit 100
        Concurrency: 20
        Iterations: 1000

    Before:

        Runtime:
            mean: 2.836 +/- 1.910
            min:  0.804
            max:  20.638
        Percentile:
            50:   2.235
            99.9: 17.204

    After:

        Runtime:
            mean: 2.448 +/- 1.039
            min:  1.702
            max:  11.048
        Percentile:
            50:   2.056
            99.9: 10.729

2. Repeat issues in the synchronized RowMerger

The repeatHandle implementation of the synchronized RowMerger doesn't
emit the rows in the same order as in the first run.

This is currently not an issue because nothing is depending on the same
order. But it will be an issue for outer joins.